requirements.txt
beautifulsoup4==4.13.5
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.76
langchain_groq==0.3.8
lxml==6.0.1
python-dotenv==1.1.1
streamlit==1.49.1
